lib: move common code from landing_session.js into session.js
