+++ lib/landing_session.js
@@ -10,18 +10,9 @@ const Session = require('./session');
 const isWindows = process.platform === 'win32';
 
 class LandingSession extends Session {
-  constructor(cli, req, dir, prid, config) {
-    super(dir, prid, config);
-    this.cli = cli;
+  constructor(cli, req, dir, prid) {
+    super(cli, dir, prid);
     this.req = req;
-    const { upstream, owner, repo } = this;
-    const upstreamHref = runSync('git', [
-      'config', '--get',
-      `remote.${upstream}.url`]).trim();
-    if (!new RegExp(`${owner}/${repo}(?:.git)?$`).test(upstreamHref)) {
-      cli.warn('Remote repository URL does not point to the expected ' +
-        `repository ${owner}/${repo}`);
-    }
   }
 
   async start(metadata) {
@@ -144,14 +135,6 @@ class LandingSession extends Session {
     await this.suggestAfterPatch(patch);
   }
 
-  getCurrentRev() {
-    return runSync('git', ['rev-parse', 'HEAD']).trim();
-  }
-
-  getCurrentBranch() {
-    return runSync('git', ['rev-parse', '--abbrev-ref', 'HEAD']).trim();
-  }
-
   async amend() {
     const { cli } = this;
     if (!this.readyToAmend()) {
@@ -207,22 +190,22 @@ class LandingSession extends Session {
       return;
     };
 
-    const notYetPushed = this.getNotYetPushedCommits();
-    const notYetPushedVerbose = this.getNotYetPushedCommits(true);
+    const stray = this.getStrayCommits();
+    const strayVerbose = this.getStrayCommits(true);
     const validateCommand = path.join(
       __dirname,
       '../node_modules/.bin/core-validate-commit' + (isWindows ? '.cmd' : '')
     );
-    await runAsync(validateCommand, notYetPushed);
+    await runAsync(validateCommand, stray);
 
     cli.separator();
     cli.log('The following commits are ready to be pushed to ' +
       `${upstream}/${branch}`);
-    cli.log(`- ${notYetPushedVerbose.join('\n- ')}`);
+    cli.log(`- ${strayVerbose.join('\n- ')}`);
     cli.separator();
 
-    let willBeLanded = notYetPushed[notYetPushed.length - 1].slice(0, 7);
-    if (notYetPushed.length > 1) {
+    let willBeLanded = stray[stray.length - 1].slice(0, 7);
+    if (stray.length > 1) {
       const head = this.getUpstreamHead().slice(0, 7);
       willBeLanded = `${head}...${willBeLanded}`;
     }
@@ -269,127 +252,6 @@ class LandingSession extends Session {
 
   async status() {
     // TODO
-  }
-
-  getNotYetPushedCommits(verbose) {
-    const { upstream, branch } = this;
-    const ref = `${upstream}/${branch}...HEAD`;
-    const gitCmd = verbose
-      ? ['log', '--oneline', '--reverse', ref] : ['rev-list', '--reverse', ref];
-    const revs = runSync('git', gitCmd).trim();
-    return revs ? revs.split('\n') : [];
-  }
-
-  getUpstreamHead(verbose) {
-    const { upstream, branch } = this;
-    return runSync('git', ['rev-parse', `${upstream}/${branch}`]).trim();
-  }
-
-  async tryAbortAm() {
-    const { cli } = this;
-    if (!this.amInProgress()) {
-      return cli.ok('No git am in progress');
-    }
-    const shouldAbortAm = await cli.prompt(
-      'Abort previous git am sessions?');
-    if (shouldAbortAm) {
-      await forceRunAsync('git', ['am', '--abort']);
-      cli.ok('Aborted previous git am sessions');
-    }
-  }
-
-  async tryAbortRebase() {
-    const { cli } = this;
-    if (!this.rebaseInProgress()) {
-      return cli.ok('No git rebase in progress');
-    }
-    const shouldAbortRebase = await cli.prompt(
-      'Abort previous git rebase sessions?');
-    if (shouldAbortRebase) {
-      await forceRunAsync('git', ['rebase', '--abort']);
-      cli.ok('Aborted previous git rebase sessions');
-    }
-  }
-
-  async tryResetHead() {
-    const { cli, upstream, branch } = this;
-    const branchName = `${upstream}/${branch}`;
-    cli.startSpinner(`Bringing ${branchName} up to date...`);
-    await runAsync('git', ['fetch', upstream, branch]);
-    cli.stopSpinner(`${branchName} is now up-to-date`);
-    const notYetPushed = this.getNotYetPushedCommits(true);
-    if (!notYetPushed.length) {
-      return;
-    }
-    cli.log(`Found stray commits in ${branchName}:\n` +
-      ` - ${notYetPushed.join('\n - ')}`);
-    const shouldReset = await cli.prompt(`Reset to ${branchName}?`);
-    if (shouldReset) {
-      await runAsync('git', ['reset', '--hard', branchName]);
-      cli.ok(`Reset to ${branchName}`);
-    }
-  }
-
-  async tryResetBranch() {
-    const { cli, upstream, branch } = this;
-    await this.tryAbortAm();
-    await this.tryAbortRebase();
-
-    const branchName = `${upstream}/${branch}`;
-    const shouldResetHead = await cli.prompt(
-      `Do you want to try reset the local ${branch} branch to ${branchName}?`);
-    if (shouldResetHead) {
-      await this.tryResetHead();
-    }
-  }
-
-  warnForMissing() {
-    const { upstream, branch, cli } = this;
-    const missing = !upstream || !branch;
-    if (!branch) {
-      cli.warn('You have not told git-node what branch you are trying' +
-               ' to land commits on.');
-      cli.separator();
-      cli.info(
-        'For example, if your want to land commits on the ' +
-        '`master` branch, you can run:\n\n' +
-        '  $ ncu-config set branch master');
-      cli.separator();
-    }
-    if (!upstream) {
-      cli.warn('You have not told git-node the remote you want to sync with.');
-      cli.separator();
-      cli.info(
-        'For example, if your remote pointing to nodejs/node is' +
-        ' `remote-upstream`, you can run:\n\n' +
-        '  $ ncu-config set upstream remote-upstream');
-      cli.separator();
-    }
-    return missing;
-  }
-
-  warnForWrongBranch() {
-    const { branch, cli } = this;
-    let rev = this.getCurrentBranch();
-    if (rev === 'HEAD') {
-      cli.warn(
-        'You are in detached HEAD state. Please run git-node on a valid ' +
-        'branch');
-      return true;
-    }
-    if (rev === branch) {
-      return false;
-    }
-    cli.warn(
-      `You are running git-node-land on \`${rev}\`,\n   but you have` +
-      ` configured \`${branch}\` to be the branch to land commits.`);
-    cli.separator();
-    cli.info(
-      `You can switch to \`${branch}\` with \`git checkout ${branch}\`, or\n` +
-      `   reconfigure the target branch with:\n\n` +
-      `  $ ncu-config set branch ${rev}`);
-    cli.separator();
-    return true;
   }
 }
