+++ lib/session.js
@@ -5,15 +5,30 @@ const fs = require('fs');
 const { getMergedConfig, getNcuDir } = require('./config');
 const rimraf = require('rimraf');
 const { readJson, writeJson, readFile, writeFile } = require('./file');
+const {
+  runAsync, runSync, forceRunAsync
+} = require('./run');
+
 const APPLYING = 'APPLYING';
 const STARTED = 'STARTED';
 const AMENDING = 'AMENDING';
 
 class Session {
-  constructor(dir, prid, config) {
+  constructor(cli, dir, prid) {
+    this.cli = cli;
     this.dir = dir;
     this.prid = prid;
-    this.config = config || getMergedConfig(this.dir);
+    this.config = getMergedConfig(this.dir);
+
+    const { upstream, owner, repo } = this;
+
+    const upstreamHref = runSync('git', [
+      'config', '--get',
+      `remote.${upstream}.url`]).trim();
+    if (!new RegExp(`${owner}/${repo}(?:.git)?$`).test(upstreamHref)) {
+      cli.warn('Remote repository URL does not point to the expected ' +
+        `repository ${owner}/${repo}`);
+    }
   }
 
   get session() {
@@ -211,6 +226,135 @@ class Session {
       this.config = sess.config;
     }
     return this;
+  }
+
+  async tryAbortAm() {
+    const { cli } = this;
+    if (!this.amInProgress()) {
+      return cli.ok('No git am in progress');
+    }
+    const shouldAbortAm = await cli.prompt(
+      'Abort previous git am sessions?');
+    if (shouldAbortAm) {
+      await forceRunAsync('git', ['am', '--abort']);
+      cli.ok('Aborted previous git am sessions');
+    }
+  }
+
+  async tryAbortRebase() {
+    const { cli } = this;
+    if (!this.rebaseInProgress()) {
+      return cli.ok('No git rebase in progress');
+    }
+    const shouldAbortRebase = await cli.prompt(
+      'Abort previous git rebase sessions?');
+    if (shouldAbortRebase) {
+      await forceRunAsync('git', ['rebase', '--abort']);
+      cli.ok('Aborted previous git rebase sessions');
+    }
+  }
+
+  async tryResetBranch() {
+    const { cli, upstream, branch } = this;
+    await this.tryAbortAm();
+    await this.tryAbortRebase();
+
+    const branchName = `${upstream}/${branch}`;
+    const shouldResetHead = await cli.prompt(
+      `Do you want to try reset the local ${branch} branch to ${branchName}?`);
+    if (shouldResetHead) {
+      await this.tryResetHead();
+    }
+  }
+
+  getCurrentRev() {
+    return runSync('git', ['rev-parse', 'HEAD']).trim();
+  }
+
+  getCurrentBranch() {
+    return runSync('git', ['rev-parse', '--abbrev-ref', 'HEAD']).trim();
+  }
+
+  getUpstreamHead() {
+    const { upstream, branch } = this;
+    return runSync('git', ['rev-parse', `${upstream}/${branch}`]).trim();
+  }
+
+  getStrayCommits(verbose) {
+    const { upstream, branch } = this;
+    const ref = `${upstream}/${branch}...HEAD`;
+    const gitCmd = verbose
+      ? ['log', '--oneline', '--reverse', ref] : ['rev-list', '--reverse', ref];
+    const revs = runSync('git', gitCmd).trim();
+    return revs ? revs.split('\n') : [];
+  }
+
+  async tryResetHead() {
+    const { cli, upstream, branch } = this;
+    const branchName = `${upstream}/${branch}`;
+    cli.startSpinner(`Bringing ${branchName} up to date...`);
+    await runAsync('git', ['fetch', upstream, branch]);
+    cli.stopSpinner(`${branchName} is now up-to-date`);
+    const stray = this.getStrayCommits(true);
+    if (!stray.length) {
+      return;
+    }
+    cli.log(`Found stray commits in ${branchName}:\n` +
+      ` - ${stray.join('\n - ')}`);
+    const shouldReset = await cli.prompt(`Reset to ${branchName}?`);
+    if (shouldReset) {
+      await runAsync('git', ['reset', '--hard', branchName]);
+      cli.ok(`Reset to ${branchName}`);
+    }
+  }
+
+  warnForMissing() {
+    const { upstream, branch, cli } = this;
+    const missing = !upstream || !branch;
+    if (!branch) {
+      cli.warn('You have not told git-node what branch you are trying' +
+               ' to land commits on.');
+      cli.separator();
+      cli.info(
+        'For example, if your want to land commits on the ' +
+        '`master` branch, you can run:\n\n' +
+        '  $ ncu-config set branch master');
+      cli.separator();
+    }
+    if (!upstream) {
+      cli.warn('You have not told git-node the remote you want to sync with.');
+      cli.separator();
+      cli.info(
+        'For example, if your remote pointing to nodejs/node is' +
+        ' `remote-upstream`, you can run:\n\n' +
+        '  $ ncu-config set upstream remote-upstream');
+      cli.separator();
+    }
+    return missing;
+  }
+
+  warnForWrongBranch() {
+    const { branch, cli } = this;
+    let rev = this.getCurrentBranch();
+    if (rev === 'HEAD') {
+      cli.warn(
+        'You are in detached HEAD state. Please run git-node on a valid ' +
+        'branch');
+      return true;
+    }
+    if (rev === branch) {
+      return false;
+    }
+    cli.warn(
+      `You are running git-node-land on \`${rev}\`,\n   but you have` +
+      ` configured \`${branch}\` to be the branch to land commits.`);
+    cli.separator();
+    cli.info(
+      `You can switch to \`${branch}\` with \`git checkout ${branch}\`, or\n` +
+      `   reconfigure the target branch with:\n\n` +
+      `  $ ncu-config set branch ${rev}`);
+    cli.separator();
+    return true;
   }
 }
